% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3vennR.R
\name{d3vennR}
\alias{d3vennR}
\title{htmlwidget to create Venn diagrams with d3.js and venn.js}
\usage{
d3vennR(data = NULL, padding = NULL, colours = NULL, fontSize = NULL,
  duration = NULL, layoutFunction = NULL, tasks = NULL, width = NULL,
  height = NULL)
}
\description{
<Add Description>
}

