% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/d3vennR.R
\name{renderD3vennR}
\alias{renderD3vennR}
\title{Widget render function for use in Shiny}
\usage{
renderD3vennR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

