% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_listener.R
\name{venn_listener}
\alias{venn_listener}
\title{Event Listeners for d3vennR.}
\usage{
venn_listener(venn, input_id, eventType = "click")
}
\arguments{
\item{venn}{A \code{d3vennR} object.}

\item{input_id}{The input id, which should be changed by this listener}

\item{eventType}{The javascript event type (e.g. "click")}
}
\description{
Event Listeners for d3vennR.
}
