% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_tooltip.R
\name{venn_tooltip}
\alias{venn_tooltip}
\title{Tooltips for d3vennR.}
\usage{
venn_tooltip(venn)
}
\arguments{
\item{venn}{A \code{d3vennR} object.}
}
\description{
\code{venn_tooltip} is a wrapper for d3vennR which adds a tooltip to the diagram with the size over selected regions shown.
}
\section{Warning}{
 This is a wrapper function that should eventually be moved to an internal option of d3
}

